Updated to SampleInheritance.cpp
getMethods and setMethods are detailed conceptualized. Special attention should be given to return type of functions. Also, how to pass the array and utilize them is key points which should be noted. Input from user and switch case etc. changes are to be made in order to run it in user friendly way.